Refact : base64toblob
